pyproject.toml
[tool.poetry]
name = "set-python-final-project"
version = "0.1.0"
description = ""
authors = ["Sergiy Dukhovenko <dukh.sergiy@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.9"
fastapi = { extras = ["standard"], version = "^0.115.6" }
uvicorn = { extras = ["standard"], version = "^0.32.1" }
yt-dlp = "^2024.7.16"
openai-whisper = { git = "https://github.com/openai/whisper.git" }


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
